Javascript Floats are double precision, so the correct corresponding Swift type should be Float, not Double.

+++ src/swift/types.js
@@ -22,7 +22,7 @@ import {
 const builtInScalarMap = {
   [GraphQLString.name]: 'String',
   [GraphQLInt.name]: 'Int',
-  [GraphQLFloat.name]: 'Float',
+  [GraphQLFloat.name]: 'Double',
   [GraphQLBoolean.name]: 'Bool',
   [GraphQLID.name]: 'GraphQLID',
 }
